+++ tests/qasm3/resources/gates.py
@@ -24,6 +24,7 @@ import os
 import pytest
 
 from pyqasm.maps.gates import (
+    FIVE_QUBIT_OP_MAP,
     FOUR_QUBIT_OP_MAP,
     ONE_QUBIT_OP_MAP,
     ONE_QUBIT_ROTATION_MAP,
@@ -42,6 +43,7 @@ VALID_GATE_NAMES = set(
     | ONE_QUBIT_ROTATION_MAP.keys()
     | THREE_QUBIT_OP_MAP.keys()
     | FOUR_QUBIT_OP_MAP.keys()
+    | FIVE_QUBIT_OP_MAP.keys()
 )
 
 
@@ -182,6 +184,30 @@ for gate in FOUR_QUBIT_OP_MAP:
     locals()[name] = _generate_four_qubit_fixture(gate)
 
 
+def _generate_five_qubit_fixture(gate_name: str):
+    @pytest.fixture()
+    def test_fixture():
+        if gate_name not in VALID_GATE_NAMES:
+            raise ValueError(f"Unknown qasm3 gate {gate_name}")
+        qasm3_string = f"""
+        OPENQASM 3;
+        include "stdgates.inc";
+
+        qubit[5] q;
+        {gate_name} q[0], q[1], q[2], q[3], q[4];
+        {gate_name} q;
+        """
+        return qasm3_string
+
+    return test_fixture
+
+
+# Generate five-qubit gate fixtures
+for gate in FIVE_QUBIT_OP_MAP:
+    name = _fixture_name(gate)
+    locals()[name] = _generate_five_qubit_fixture(gate)
+
+
 def _generate_custom_op_fixture(op_name: str):
     print(os.getcwd())
 
@@ -243,6 +269,8 @@ for gate in already_tested_triple_op:
     triple_op_tests.remove(_fixture_name(gate))
 
 four_op_tests = [_fixture_name(s) for s in FOUR_QUBIT_OP_MAP]
+
+five_op_tests = [_fixture_name(s) for s in FIVE_QUBIT_OP_MAP]
 
 custom_op_tests = [_fixture_name(s) for s in CUSTOM_OPS]